fixed the backing for map and ISBN
