+++ code/app/src/main/java/com/example/booker/data/RequestList.java
@@ -2,13 +2,12 @@ package com.example.booker.data;
 
 import android.content.Context;
 import android.content.Intent;
+import android.os.Parcelable;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
-import android.widget.Button;
-import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
@@ -16,13 +15,11 @@ import androidx.annotation.Nullable;
 
 import com.example.booker.R;
 import com.example.booker.activities.MapsActivity;
-import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FieldValue;
@@ -30,10 +27,8 @@ import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 
+import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.List;
-
-import static androidx.constraintlayout.motion.widget.Debug.getLocation;
 
 /**
  * RequestList class is the adapter for request list, the function of it
@@ -42,7 +37,7 @@ import static androidx.constraintlayout.motion.widget.Debug.getLocation;
  */
 public class RequestList extends ArrayAdapter<Request> {
     final String TAG = "request update tag";
-    private ArrayList<Request> requests;
+    public ArrayList<Request> requests;
     private Context context;
     private FirebaseAuth mAuth;
     private FirebaseFirestore db;
@@ -93,136 +88,163 @@ public class RequestList extends ArrayAdapter<Request> {
         request_accept.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                //find the path to the correspond Request
-                final DocumentReference documentRef = db.collection("User")
-                        .document(mAuth.getCurrentUser().getUid())
-                        .collection("Lend")
-                        .document(request.getBook_name());
-                DocumentReference owner_path = db.collection("User")
-                        .document(mAuth.getCurrentUser().getUid())
-                        .collection("Lend")
-                        .document(request.getBook_name());
-                //change book status for owner
-                owner_path.update("status","accepted");
-                owner_path.update("requests", FieldValue.arrayRemove());
 
-                //change borrow status for accepted user
-                db.collection("User").get()
-                        .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                            @Override
-                            public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                                if(task.isSuccessful()){
-                                    for (QueryDocumentSnapshot document : task.getResult()) {
-                                        //find a matched name, update the status of its borrowed book
-                                        if(document.getId().equals(request.getUser_name())){
-                                            //change status to accepted
-                                            db.collection("User").document(document.getId()).collection("Borrowed")
-                                                    .document(request.getBook_name()).update("status","accepted")
-                                                    .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                                        @Override
-                                                        public void onSuccess(Void aVoid) {
-                                                            Log.d(TAG, "Correspond user accept status successfully updated!");
-                                                        }
-                                                    })
-                                                    .addOnFailureListener(new OnFailureListener() {
-                                                        @Override
-                                                        public void onFailure(@NonNull Exception e) {
-                                                            Log.d(TAG, "Correspond user accept status failed to updated!");
-                                                        }
-                                                    });
+
+                Intent mapIntent = new Intent(context, MapsActivity.class);
+
+
+                mapIntent.putExtra("bookName", request.getBook_name());
+                Log.d(TAG, "putExtra: Borrower:"+request.getBook_name());
+
+//                mapIntent.putExtra("lenderName", request.getUser_name());
+//                Log.d(TAG, "putExtra: Borrower:"+request.getUser_name());
+
+                mapIntent.putExtra("borrowerName", request.getUser_name());
+                Log.d(TAG, "putExtra: Borrower:"+request.getUser_name());
+                context.startActivity(mapIntent);
+
+
+
+
+
+
+
+
+
+
+
+//                //find the path to the correspond Request
+//                final DocumentReference documentRef = db.collection("User")
+//                        .document(mAuth.getCurrentUser().getUid())
+//                        .collection("Lend")
+//                        .document(request.getBook_name());
+//                DocumentReference owner_path = db.collection("User")
+//                        .document(mAuth.getCurrentUser().getUid())
+//                        .collection("Lend")
+//                        .document(request.getBook_name());
+//                //change book status for owner
+//                owner_path.update("status","accepted");
+//                owner_path.update("requests", FieldValue.arrayRemove());
 //
-//                                            db.collection("trans").document(document.getId())
-//                                                    .set(request.getBook_name())
+//                //change borrow status for accepted user
+//                db.collection("User").get()
+//                        .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+//                            @Override
+//                            public void onComplete(@NonNull Task<QuerySnapshot> task) {
+//                                if(task.isSuccessful()){
+//                                    for (QueryDocumentSnapshot document : task.getResult()) {
+//                                        //find a matched name, update the status of its borrowed book
+//                                        if(document.getId().equals(request.getUser_name())){
+//                                            //change status to accepted
+//                                            db.collection("User").document(document.getId()).collection("Borrowed")
+//                                                    .document(request.getBook_name()).update("status","accepted")
 //                                                    .addOnSuccessListener(new OnSuccessListener<Void>() {
 //                                                        @Override
 //                                                        public void onSuccess(Void aVoid) {
-//                                                            // These are a method which gets executed when the task is succeeded
-//                                                            Log.d(TAG, "trans: Data has been added successfully!");
+//                                                            Log.d(TAG, "Correspond user accept status successfully updated!");
 //                                                        }
 //                                                    })
 //                                                    .addOnFailureListener(new OnFailureListener() {
 //                                                        @Override
 //                                                        public void onFailure(@NonNull Exception e) {
-//                                                            // These are a method which gets executed if there’s any problem
-//                                                            Log.d(TAG, "trans: Data could not be added!" + e.toString());
+//                                                            Log.d(TAG, "Correspond user accept status failed to updated!");
 //                                                        }
 //                                                    });
+////
+////                                            db.collection("trans").document(document.getId())
+////                                                    .set(request.getBook_name())
+////                                                    .addOnSuccessListener(new OnSuccessListener<Void>() {
+////                                                        @Override
+////                                                        public void onSuccess(Void aVoid) {
+////                                                            // These are a method which gets executed when the task is succeeded
+////                                                            Log.d(TAG, "trans: Data has been added successfully!");
+////                                                        }
+////                                                    })
+////                                                    .addOnFailureListener(new OnFailureListener() {
+////                                                        @Override
+////                                                        public void onFailure(@NonNull Exception e) {
+////                                                            // These are a method which gets executed if there’s any problem
+////                                                            Log.d(TAG, "trans: Data could not be added!" + e.toString());
+////                                                        }
+////                                                    });
+//
+//
+//
+//                                        }
+//                                    }
+//                                }else{
+//                                    Log.d(TAG, "Fail to find user col/doc!");
+//                                }
+//                            }
+//                        });
+//                //update status for other users, loop all users
+//                for(final Request user_request : requests){
+//                    if(!user_request.getUser_name().equals(request.getUser_name())) {
+//                        db.collection("User").get()
+//                                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+//                                    @Override
+//                                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+//                                        if (task.isSuccessful()) {
+//                                            for (QueryDocumentSnapshot document : task.getResult()) {
+//                                                //find a matched name, update the status of its borrowed book
+//                                                if (document.getId().equals(user_request.getUser_name())) {
+//                                                    //if decline a request, then for the user trying to borrow the book, it will disappear and send notification
+//                                                    //delete the correspond book in borrower's borrowed list
+//                                                    db.collection("User").document(document.getId()).collection("Borrowed")
+//                                                            .document(user_request.getBook_name()).delete()
+//                                                            .addOnSuccessListener(new OnSuccessListener<Void>() {
+//                                                                @Override
+//                                                                public void onSuccess(Void aVoid) {
+//
+//                                                                    //notification code write here
+//                                                                    Log.d(TAG, "Correspond user accept status successfully updated!");
+//                                                                }
+//                                                            })
+//                                                            .addOnFailureListener(new OnFailureListener() {
+//                                                                @Override
+//                                                                public void onFailure(@NonNull Exception e) {
+//                                                                    Log.d(TAG, "Correspond user accept status failed to updated!");
+//                                                                }
+//                                                            });
+//                                                }
+//                                            }
+//                                        } else {
+//                                            Log.d(TAG, "Fail to find user col/doc!");
+//                                        }
+//                                    }
+//                                });
+//                    }else{//change the borrower field in owner's book
+//                        db.collection("User").document(mAuth.getCurrentUser().getUid()).collection("Lend")
+//                                .document(request.getBook_name()).update("borrower",user_request.getUser_name())
+//                                .addOnSuccessListener(new OnSuccessListener<Void>() {
+//                                    @Override
+//                                    public void onSuccess(Void aVoid) {
+//                                        Log.d(TAG, "Correspond user accept status successfully updated!");
+//                                    }
+//                                })
+//                                .addOnFailureListener(new OnFailureListener() {
+//                                    @Override
+//                                    public void onFailure(@NonNull Exception e) {
+//                                        Log.d(TAG, "Correspond user accept status failed to updated!");
+//                                    }
+//                                });
+//                    }
+//                }
+//
+//
+//                //delete all elements in Request array(update request list array) ???
+//                documentRef.update("requests",null).addOnCompleteListener(new OnCompleteListener<Void>() {
+//                    @Override
+//                    public void onComplete(@NonNull Task<Void> task) {
+//                        if(task.isSuccessful()){
+//                            Log.d(TAG, "Remove all users successfully! ");
+//                        }else{
+//                            Log.d(TAG, "Fail to remove all users!");
+//                        }
+//                    }
+//                });
 
-
-
-                                        }
-                                    }
-                                }else{
-                                    Log.d(TAG, "Fail to find user col/doc!");
-                                }
-                            }
-                        });
-                //update status for other users, loop all users
-                for(final Request user_request : requests){
-                    if(!user_request.getUser_name().equals(request.getUser_name())) {
-                        db.collection("User").get()
-                                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                                    @Override
-                                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                                        if (task.isSuccessful()) {
-                                            for (QueryDocumentSnapshot document : task.getResult()) {
-                                                //find a matched name, update the status of its borrowed book
-                                                if (document.getId().equals(user_request.getUser_name())) {
-                                                    //if decline a request, then for the user trying to borrow the book, it will disappear and send notification
-                                                    //delete the correspond book in borrower's borrowed list
-                                                    db.collection("User").document(document.getId()).collection("Borrowed")
-                                                            .document(user_request.getBook_name()).delete()
-                                                            .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                                                @Override
-                                                                public void onSuccess(Void aVoid) {
-
-                                                                    //notification code write here
-                                                                    Log.d(TAG, "Correspond user accept status successfully updated!");
-                                                                }
-                                                            })
-                                                            .addOnFailureListener(new OnFailureListener() {
-                                                                @Override
-                                                                public void onFailure(@NonNull Exception e) {
-                                                                    Log.d(TAG, "Correspond user accept status failed to updated!");
-                                                                }
-                                                            });
-                                                }
-                                            }
-                                        } else {
-                                            Log.d(TAG, "Fail to find user col/doc!");
-                                        }
-                                    }
-                                });
-                    }else{//change the borrower field in owner's book
-                        db.collection("User").document(mAuth.getCurrentUser().getUid()).collection("Lend")
-                                .document(request.getBook_name()).update("borrower",user_request.getUser_name())
-                                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                    @Override
-                                    public void onSuccess(Void aVoid) {
-                                        Log.d(TAG, "Correspond user accept status successfully updated!");
-                                    }
-                                })
-                                .addOnFailureListener(new OnFailureListener() {
-                                    @Override
-                                    public void onFailure(@NonNull Exception e) {
-                                        Log.d(TAG, "Correspond user accept status failed to updated!");
-                                    }
-                                });
-                    }
-                }
-
-
-                //delete all elements in Request array(update request list array) ???
-                documentRef.update("requests",null).addOnCompleteListener(new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        if(task.isSuccessful()){
-                            Log.d(TAG, "Remove all users successfully! ");
-                        }else{
-                            Log.d(TAG, "Fail to remove all users!");
-                        }
-                    }
-                });
+                //end
 
 
 
@@ -265,18 +287,18 @@ public class RequestList extends ArrayAdapter<Request> {
 
 
 
-                Intent mapIntent = new Intent(context, MapsActivity.class);
-
-
-                mapIntent.putExtra("bookName", request.getBook_name());
-                Log.d(TAG, "putExtra: Borrower:"+request.getBook_name());
-
-//                mapIntent.putExtra("lenderName", request.getUser_name());
+//                Intent mapIntent = new Intent(context, MapsActivity.class);
+//
+//
+//                mapIntent.putExtra("bookName", request.getBook_name());
+//                Log.d(TAG, "putExtra: Borrower:"+request.getBook_name());
+//
+////                mapIntent.putExtra("lenderName", request.getUser_name());
+////                Log.d(TAG, "putExtra: Borrower:"+request.getUser_name());
+//
+//                mapIntent.putExtra("borrowerName", request.getUser_name());
 //                Log.d(TAG, "putExtra: Borrower:"+request.getUser_name());
-
-                mapIntent.putExtra("borrowerName", request.getUser_name());
-                Log.d(TAG, "putExtra: Borrower:"+request.getUser_name());
-                context.startActivity(mapIntent);
+//                context.startActivity(mapIntent);
 
 
             }
@@ -323,16 +345,16 @@ public class RequestList extends ArrayAdapter<Request> {
                             }
                         });
                 //delete the correspond item in array
-                documentRef.update("requests",FieldValue.arrayRemove(request.getUser_name())).addOnCompleteListener(new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull  Task<Void> task) {
-                        if(task.isSuccessful()){
-                            Log.d(TAG, "Remove single user successfully! ");
-                        }else{
-                            Log.d(TAG, "Fail to remove single user!");
-                        }
-                    }
-                });
+//                documentRef.update("requests",FieldValue.arrayRemove(request.getUser_name())).addOnCompleteListener(new OnCompleteListener<Void>() {
+//                    @Override
+//                    public void onComplete(@NonNull  Task<Void> task) {
+//                        if(task.isSuccessful()){
+//                            Log.d(TAG, "Remove single user successfully! ");
+//                        }else{
+//                            Log.d(TAG, "Fail to remove single user!");
+//                        }
+//                    }
+//                });
             }
         });
